Extracted logic into AbstractMainPlugin
