+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Plugin.php
@@ -12,19 +12,17 @@ use GraphQLAPI\GraphQLAPI\HybridServices\ModuleResolvers\PluginManagementFunctio
 use GraphQLAPI\GraphQLAPI\PluginConfiguration;
 use GraphQLAPI\GraphQLAPI\Registries\ModuleRegistry;
 use GraphQLAPI\GraphQLAPI\Security\UserAuthorization;
-use GraphQLAPI\GraphQLAPI\Services\CustomPostTypes\AbstractCustomPostType;
 use GraphQLAPI\GraphQLAPI\Services\Helpers\EndpointHelpers;
 use GraphQLAPI\GraphQLAPI\Services\Helpers\MenuPageHelper;
 use GraphQLAPI\GraphQLAPI\Services\MenuPages\AboutMenuPage;
 use GraphQLAPI\GraphQLAPI\Services\MenuPages\ModulesMenuPage;
 use GraphQLAPI\GraphQLAPI\Services\MenuPages\SettingsMenuPage;
 use GraphQLAPI\GraphQLAPI\Services\Menus\Menu;
-use GraphQLAPI\PluginSkeleton\AbstractPlugin;
+use GraphQLAPI\PluginSkeleton\AbstractMainPlugin;
 use PoP\ComponentModel\Facades\Instances\InstanceManagerFacade;
 use PoP\Engine\AppLoader;
-use PoP\Root\Container\ContainerBuilderUtils;
 
-class Plugin extends AbstractPlugin
+class Plugin extends AbstractMainPlugin
 {
     /**
      * Plugin's namespace
@@ -39,29 +37,11 @@ class Plugin extends AbstractPlugin
 
     /**
      * Plugin set-up, executed immediately when loading the plugin.
-     * There are three stages for this plugin, and for each extension plugin:
-     * `setup`, `initialize` and `boot`.
-     *
-     * This is because:
-     *
-     * - The plugin must execute its logic before the extensions
-     * - The services can't be booted before all services have been initialized
-     *
-     * To attain the needed order, we execute them using hook "plugins_loaded":
-     *
-     * 1. GraphQL API => setup(): immediately
-     * 2. GraphQL API extensions => setup(): priority 0
-     * 3. GraphQL API => initialize(): priority 10
-     * 4. GraphQL API extensions => initialize(): priority 20
-     * 5. GraphQL API => bootSystem(): priority 30
-     * 3. GraphQL API => configure(): priority 40
-     * 4. GraphQL API extensions => configure(): priority 50
-     * 5. GraphQL API => bootApplication(): priority 60
-     * 6. GraphQL API => boot(): priority 70
-     * 7. GraphQL API extensions => boot(): priority 80
      */
     public function setup(): void
     {
+        parent::setup();
+
         // Functions to execute when activating/deactivating the plugin
         \register_deactivation_hook($this->getPluginFile(), [$this, 'deactivate']);
         /**
@@ -140,27 +120,6 @@ class Plugin extends AbstractPlugin
                 $this->showReleaseNotesInAdminNotice();
             }
         );
-
-        /**
-         * Wait until "plugins_loaded" to initialize the plugin, because:
-         *
-         * - ModuleListTableAction requires `wp_verify_nonce`, loaded in pluggable.php
-         * - Allow other plugins to inject their own functionality
-         */
-        \add_action('plugins_loaded', [$this, 'initialize'], 10);
-        \add_action('plugins_loaded', function () {
-            \do_action(self::HOOK_INITIALIZE_EXTENSION_PLUGIN);
-        }, 20);
-        \add_action('plugins_loaded', [$this, 'bootSystem'], 30);
-        \add_action('plugins_loaded', [$this, 'configure'], 40);
-        \add_action('plugins_loaded', function () {
-            \do_action(self::HOOK_CONFIGURE_EXTENSION_PLUGIN);
-        }, 50);
-        \add_action('plugins_loaded', [$this, 'bootApplication'], 60);
-        \add_action('plugins_loaded', [$this, 'boot'], 70);
-        \add_action('plugins_loaded', function () {
-            \do_action(self::HOOK_BOOT_EXTENSION_PLUGIN);
-        }, 80);
     }
 
     /**
@@ -326,11 +285,6 @@ class Plugin extends AbstractPlugin
         AppLoader::bootSystem(
             ...PluginConfiguration::getContainerCacheConfiguration()
         );
-    }
-
-    public function boot(): void
-    {
-        // Doing nothing yet
     }
 
     /**
